+++ src/App.js
@@ -1,98 +1,28 @@
-import { DivOverlay } from 'leaflet';
-import React, { useEffect, useRef, useState } from 'react';
-import { MapContainer, TileLayer } from 'react-leaflet';
-import MarkerCluster from './components/MarkerCluster';
+import React from 'react';
+
+import { Route, Routes } from 'react-router-dom';
+import NavBar from './components/NavBar';
+import Home from './pages/Home'
+import About from './pages/About'
+import Contact from './pages/Contact'
 
 function App() {
 
-  const [data, setData] = useState([])
-  const [search, setSearch] = useState('')
-  const [searchW, setSearchW] = useState(null)
-  const [autocomplete, setAutocomplete] = useState([])
-  const [numberOfEntities, setNumberOfEntities] = useState(0)
-  const [loadingSearch, setLoadingSearch] = useState(false)
-  const [autocompleteLoading, setAutocompleteLoading] = useState(false)
 
-  const searchRef = useRef(null)
-
-  const fetchData = () => {
-    fetch(`data/${searchRef.current.value}`)
-      .then(res => res.json())
-      .then(data =>{
-        setData(data);
-        setNumberOfEntities(data.results.length);
-        setSearchW(searchRef.current.value)
-        setLoadingSearch(false)
-      })
-      .catch(error => console.log("Fetch data error: " + error))
-  }
-
-  function onClickButton() {
-    setLoadingSearch(true)
-    fetchData()
-  }
-
-useEffect(() => {
-  setAutocompleteLoading(true)
-  fetch(`autocomplete/${search}`)
-  .then(res => res.json())
-  .then(data => {
-      setAutocomplete(data.types)
-    })
-  setAutocompleteLoading(false)
-}, [search])
   
 
   return (
-    <div className="App container">
-      <h3 className='col fs-3'>Wikidata Atlas <i class="bi bi-pin-map-fill"></i></h3>
-      <div className='row my-2 justify-content-end'>
-      
-        <input className='col-3 me-1 border rounded text-dark' 
-          ref={searchRef}
-          onChange={() => setSearch(searchRef.current.value)} 
-          placeholder='Search Wikidata'
-          list='list'
-        />
-        {
-          autocompleteLoading ? <div className='bg-danger'>Loading...</div> :
-        <datalist id='list'>
-        {
-          autocomplete.map(e=>{
-            var values = e[1]
-            return <option value={values.label}>{values.description}</option>
-          })
-        }
-        </datalist>
-        }
-        <button className='btn btn-info col-1' onClick={onClickButton}>
-          {
-            loadingSearch ? <span class="spinner-border spinner-border-sm" role="status" aria-hidden="true"></span> : <i className="bi bi-search"></i>
-          }
-        </button>
+      <>
+      <NavBar />
+      <div className='cotainer'>
+        <Routes>
+          <Route path='/' element={<Home />} />
+          <Route path='/about' element={<About />} />
+          <Route path='/contact' element={<Contact />} />
+        </Routes>
       </div>
-
-      {
-        // (numberOfEntities > 0) && <div>HOLA</div>
-        numberOfEntities > 0 ? 
-        <div className='col text-success h-50'>
-          <b> {numberOfEntities} </b> entities instance of <b>{searchW}</b> founded.
-        </div> : <div className='col text-secondary h-50'>Search an entity type (P31 property)...</div>
-      }
-
       
-      <MapContainer center={[0,0]} zoom={2} minZoom={2} maxZoom={18} className='map'>
-             
-        <TileLayer
-          url='https://{s}.tile.openstreetmap.fr/hot/{z}/{x}/{y}.png'
-          attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors, Tiles style by <a href="https://www.hotosm.org/" target="_blank">Humanitarian OpenStreetMap Team</a> hosted by <a href="https://openstreetmap.fr/" target="_blank">OpenStreetMap France</a>'
-        />
-
-
-        
-        <MarkerCluster markers={data} />
-      </MapContainer>
-    </div>
+      </>
   );
 }
