+++ src/pages/About.jsx
@@ -1,0 +1,7 @@
+function About() {
+    return ( 
+        <h1>About</h1>
+     );
+}
+
+export default About;
