+++ src/pages/Contact.jsx
@@ -1,0 +1,7 @@
+function Contact() {
+    return ( 
+        <h1>Contact</h1>
+     );
+}
+
+export default Contact;
